Asset, Transaction value 자료형 Long으로 수정. close#128, close#130 (#132)
* Asset, Transaction value 자료형 Long으로 수정. close#128, close#130

* 일부 테스트 ì 알맞게 추가 수정.

+++ backend/src/main/java/com/back/domain/asset/Dto/AssetDto.java
@@ -9,7 +9,7 @@ public record AssetDto (
         int memberId,
         String name,
         String assetType,
-        int assetValue,
+        Long assetValue,
         LocalDateTime createDate,
         LocalDateTime modifyDate
 ) {
