+++ backend/src/main/java/com/back/domain/transactions/dto/AccountTransactionDto.java
@@ -8,7 +8,7 @@ public record AccountTransactionDto(
         int id,
         int accountId,
         String type,
-        int amount,
+        Long amount,
         String content,
         LocalDateTime date,
         LocalDateTime createDate,
