+++ backend/src/main/java/com/back/domain/transactions/dto/CreateAccTracRequestDto.java
@@ -3,11 +3,11 @@ package com.back.domain.transactions.dto;
 public record CreateAccTracRequestDto(
         int accountId,
         String type,
-        int amount,
+        Long amount,
         String content,
         String date
 ) {
-    public CreateAccTracRequestDto(int accountId, String type, int amount, String content, String date) {
+    public CreateAccTracRequestDto(int accountId, String type, Long amount, String content, String date) {
         this.accountId = accountId;
         this.type = type;
         this.amount = amount;
