+++ backend/src/main/java/com/back/domain/transactions/entity/AccountTransaction.java
@@ -24,7 +24,7 @@ public class AccountTransaction extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private TransactionType type; // 거래 유형
 
-    private int amount; // 거래량
+    private Long amount; // 거래량
     private String content; // 필요 시 메모
     private LocalDateTime date; // 체결일
 }
